refactor: rename useAgentConversation => useConversationWith
